feat: async request from admin, delete a product from client side js

+++ controllers/admin.js
@@ -175,16 +175,19 @@ exports.getAdminProducts = (req, res, next) => {
             return next(error);
         });
 }
-exports.postDeleteProduct = (req, res, next) => {
-    const prodId = req.body.id;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+exports.deleteProduct = (req, res, next) => {
+    const prodId = req.params.id;
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found.'))
+            }
+            return Product.deleteOne({ _id: prodId, userId: req.user._id })
+        })
         .then(() => {
-            console.log('Destroyed product');
-            res.redirect('/admin/products')
+            res.status(200).json({ message: 'Success!' });
         })
         .catch(err => {
-            const error = new Error(err);
-            error.httpStatusCode = 500;
-            return next(error);
+            res.status(500).json({ message: 'Deleting product failed.' });
         });
 }
